Move statistics to session statistics component
