+++ src/app/app-routing.module.ts
@@ -16,6 +16,7 @@ import { DashboardsComponent } from './dashboards/dashboards.component';
 import { StepComponent } from './step/step-component/step.component';
 import { TerminalComponent } from './step/terminal/terminal.component';
 import { RolesComponent } from './configuration/roles/roles/roles.component';
+import { SessionStatisticsComponent } from './session-statistics/session-statistics.component';
 
 const routes: Routes = [
   {path: '', redirectTo: '/home', pathMatch: 'full'},
@@ -25,6 +26,12 @@ const routes: Routes = [
     component: HomeComponent,
     canActivate: [
       AuthGuard
+    ],
+    children: [
+      {
+        path: 'statistics/sessions',
+        component: SessionStatisticsComponent
+      }
     ]
   },
   {
