+++ src/app/home/home.component.html
@@ -1,105 +1,16 @@
 <div class="main-container">
     <header class="header header-7" app-header></header>
-    <p class="welcome-msg">Welcome to the Admin Page</p>
-
-    <div class="clr-row" *rbac="['progresses.list']">
-      <div class="clr-col-lg-6 clr-col-md-8 clr-col-12">
-          <div class="card">
-              <h3 class="card-header">Usage Statistics For Active Events</h3>
-              <div class="card-block">
-                  <h4 class="card-title">Started User Sessions</h4>
-                  <div class="card-text">
-                    <form clrForm [formGroup]="chartDetails" [class.clr-error]="chartDetails.errors?.endDateLowerThanStartDate">
-                      <select formControlName="observationPeriod" >
-                        <option value="daily">Daily</option>
-                        <option value="weekly">Weekly</option>
-                        <option value="monthly">Monthly</option>
-                      </select>
-                      <clr-input-container>
-                        <label class="clr-col-lg-2 clr-col-md-2 clr-col-4">Start Date</label>
-                        <input
-                          class="clr-col-lg-2 clr-col-md-2 clr-col-4"
-                          clrInput
-                          type="text"
-                          name="startDate"
-                          formControlName="startDate"
-                          required
-                          readonly
-                        />
-                        <clr-control-error *clrIfError="'required'">
-                          Start Date required.
-                        </clr-control-error>
-                        <clr-control-error *clrIfError="'startDateIsInFuture'">
-                          Start Date should be today or in the past.
-                        </clr-control-error>
-                      </clr-input-container>
-                      <clr-input-container>
-                        <label class="clr-col-lg-2 clr-col-md-2 clr-col-4">End Date</label>
-                        <input
-                          class="clr-col-lg-2 clr-col-md-2 clr-col-4"
-                          clrInput
-                          type="text"
-                          name="endDate"
-                          formControlName="endDate"
-                          required
-                          readonly
-                        />
-                        <clr-control-error *clrIfError="'required'">
-                          End Date required.
-                        </clr-control-error>
-                      </clr-input-container>
-                      <span class="clr-subtext" *ngIf="chartDetails.errors?.endDateLowerThanStartDate">
-                        <clr-icon class="clr-validate-icon" shape="exclamation-circle"></clr-icon> Start date must occur before end time!
-                      </span>
-                    </form>
-                  </div>
-                  <div style="display: block">
-                    <canvas baseChart
-                        [data]="barChartData"
-                        [options]="barChartOptions"
-                        [plugins]="barChartPlugins"
-                        [type]="barChartType"
-                        (chartHover)="chartHovered($event)"
-                        (chartClick)="chartClicked($event)">
-                    </canvas>
-                </div>
-                <div class="card-footer">
-                  <clr-signpost>
-                    <button class="btn btn-sm btn-link" clrSignpostTrigger>Change Start Date</button>
-                    <clr-signpost-content [clrPosition]="'top-right'" *clrIfOpen #startDateSignpost>
-                      <dl-date-time-picker (change)="setStartDate($event)" [startView]="startView" maxView="year" [minView]="minView">
-                      </dl-date-time-picker>
-                    </clr-signpost-content>
-                  </clr-signpost>
-                  <clr-signpost>
-                    <button class="btn btn-link" clrSignpostTrigger>Change End Date</button>
-                    <clr-signpost-content [clrPosition]="'top-right'" *clrIfOpen #endDateSignpost>
-                      <dl-date-time-picker (change)="setEndDate($event)" [startView]="startView" maxView="year" [minView]="minView">
-                      </dl-date-time-picker>
-                    </clr-signpost-content>
-                  </clr-signpost>
-                </div>
-              </div>
-          </div>
-      </div>
-      <div class="clr-col-lg-6 clr-col-md-8 clr-col-12">
-        <div class="card">
-          <h3 class="card-header">Overview Scenario Sessions</h3>
-          <div class="card-block">
-            <h4 class="card-title">Total Sessions</h4>
-            <div class="card-text">
-              <clr-datagrid>
-                <clr-dg-column [clrDgField]="'key'" [clrDgSortOrder]="ascSort">Scenario</clr-dg-column>
-                <clr-dg-column>Total Sessions</clr-dg-column>
-            
-                <clr-dg-row *clrDgItems="let item of totalSessionsPerScenario | keyvalue">
-                    <clr-dg-cell>{{item.key}}</clr-dg-cell>
-                    <clr-dg-cell>{{item.value}}</clr-dg-cell>
-                </clr-dg-row>
-            </clr-datagrid>
-            </div>
-          </div>
+    <!-- <p class="welcome-msg">Welcome to the Admin Page</p> -->
+    <div class="content-container">
+        <nav class="sidenav">
+            <section class="sidenav-content">
+                <a href="javascript://" [routerLink]="['statistics/sessions']" [routerLinkActive]="'active'" class="nav-link" *rbac="['progresses.list', 'scenarios.list']; op 'AND'">
+                    Session Statistics
+                </a>
+            </section>
+        </nav>
+        <div class="content-area">
+            <router-outlet></router-outlet>
         </div>
-      </div>
     </div>
 </div>
