+++ src/app/home/home.component.scss
@@ -1,7 +1,4 @@
-.progress-dashboard {
-  height: 100%;
-}
 .welcome-msg {
-  display: flex;
-  align-self: center;
-}
+    display: flex;
+    align-self: center;
+  }
