+++ src/app/session-statistics/session-statistics.component.html
@@ -1,0 +1,90 @@
+<div class="main-container" *rbac="['progresses.list']">
+  <h3>Session Statistics For Active Events</h3>
+  <form [formGroup]="chartDetails" [class.clr-error]="chartDetails.errors?.endDateLowerThanStartDate">
+    <select clrSelect formControlName="observationPeriod" >
+      <option value="daily">Daily</option>
+      <option value="weekly">Weekly</option>
+      <option value="monthly">Monthly</option>
+    </select>
+    <div class="clr-row">
+      <div class="clr-col-lg-4 clr-col-md-4 clr-col-12">
+        <clr-input-container class="date-label-input">
+          <label>Start Date</label>
+          <input
+            clrInput
+            type="text"
+            name="startDate"
+            formControlName="startDate"
+            required
+            readonly
+          />
+          <clr-control-error *clrIfError="'required'">
+            Start Date required.
+          </clr-control-error>
+          <clr-control-error *clrIfError="'startDateIsInFuture'">
+            Start Date should be today or in the past.
+          </clr-control-error>
+        </clr-input-container>
+        <clr-signpost class="signpost">
+          <button class="btn btn-sm btn-link" clrSignpostTrigger>Change Start Date</button>
+          <clr-signpost-content [clrPosition]="'top-right'" *clrIfOpen #startDateSignpost>
+            <dl-date-time-picker (change)="setStartDate($event)" [startView]="startView" maxView="year" [minView]="minView">
+            </dl-date-time-picker>
+          </clr-signpost-content>
+        </clr-signpost>
+      </div>
+      <div class="clr-col-lg-4 clr-col-md-4 clr-col-12">
+        <clr-input-container class="date-label-input">
+          <label>End Date</label>
+          <input
+            clrInput
+            type="text"
+            name="endDate"
+            formControlName="endDate"
+            required
+            readonly
+          />
+          <clr-control-error *clrIfError="'required'">
+            End Date required.
+          </clr-control-error>
+        </clr-input-container>
+        <clr-signpost class="signpost">
+          <button class="btn btn-link" clrSignpostTrigger>Change End Date</button>
+          <clr-signpost-content [clrPosition]="'top-right'" *clrIfOpen #endDateSignpost>
+            <dl-date-time-picker (change)="setEndDate($event)" [startView]="startView" maxView="year" [minView]="minView">
+            </dl-date-time-picker>
+          </clr-signpost-content>
+        </clr-signpost>
+        <span class="clr-subtext" *ngIf="chartDetails.errors?.endDateLowerThanStartDate">
+          <clr-icon class="clr-validate-icon" shape="exclamation-circle"></clr-icon> Start date must occur before end time!
+        </span>
+      </div>
+    </div>
+  </form>
+  <div class="clr-row">
+    <div class="clr-col-lg-8 clr-col-md-8 clr-col-12">
+      <h4>Started Sessions (Detailed)</h4>
+      <div style="display: block">
+        <canvas baseChart
+          [data]="barChartData"
+          [options]="barChartOptions"
+          [plugins]="barChartPlugins"
+          [type]="barChartType"
+          (chartHover)="chartHovered($event)"
+          (chartClick)="chartClicked($event)">
+        </canvas>
+      </div>
+    </div>
+    <div class="clr-col-lg-4 clr-col-md-4 clr-col-12">
+      <h4>Started Sessions (Overview)</h4>
+      <clr-datagrid>
+        <clr-dg-column [clrDgField]="'key'" [clrDgSortOrder]="ascSort">Scenario</clr-dg-column>
+        <clr-dg-column>Total Sessions</clr-dg-column>
+        <clr-dg-row *clrDgItems="let item of totalSessionsPerScenario | keyvalue">
+          <clr-dg-cell>{{item.key}}</clr-dg-cell>
+          <clr-dg-cell>{{item.value}}</clr-dg-cell>
+        </clr-dg-row>
+      </clr-datagrid>
+    </div>
+  </div>
+</div>
