+++ src/app/session-statistics/session-statistics.component.scss
@@ -1,0 +1,8 @@
+.date-label-input {
+  align-items: center;
+  flex-direction: row;
+}
+
+.signpost {
+  margin-left: -0.6rem;
+}
